requirements.txt
# Streamlit Cloud (light — no torch/Ollama)
streamlit>=1.28.0
python-dotenv>=1.0.0
PyPDF2>=3.0.0
google-generativeai>=0.8.0
gTTS>=2.5.0
